+++ packages/orq-rc/RELEASES.md
@@ -2289,3 +2289,13 @@ Based on:
 - [typescript v3.8.0-rc.35] packages/orq-rc
 ### Releases
 - [NPM v3.8.0-rc.35] https://www.npmjs.com/package/@orq-ai/node/v/3.8.0-rc.35 - packages/orq-rc
+
+## 2025-06-18 14:17:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.565.0 (2.630.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.8.0-rc.36] packages/orq-rc
+### Releases
+- [NPM v3.8.0-rc.36] https://www.npmjs.com/package/@orq-ai/node/v/3.8.0-rc.36 - packages/orq-rc